move rpt-1.5 to a dedicated package, fix tests
